platform: nrfx_glue: Use assert() to implement NRFX_ASSERT()
assert() is used in the source tree much wider than TFM_ASSERT(),
so it seems to be a better fit for implementation of NRFX_ASSERT().
Furthermore, TFM_ASSERT() calls tfm_log_printf(), which is not
available in BL2, where we also need nrfx drivers (underneath
the CMSIS ones) and thus we need proper NRFX_ASSERT() there.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>