tfm_common_s.ld: Rearrange SG section to accomodate the nRF IDAU
Nordic Semiconductor's IDAU "System Protection Unit" has restrictions
on the placement of the non-secure callable area. It must be placed at
the end of one of its regions. Since the MCUboot trailer must be placed
after the image, but within the secure area, the NSC area should not be
at the end of the last secure region, to not waste an entire region for
the trailer.

This patch places the .gnu.sgstubs section right before the text section
and allows it to be placed in the FLASH memory region as opposed to its
own VENEERS region. Also, this modifies the alignment logic because of
the same placement restrictions.

For any other platform, the change should not have any effect as long as
TFM_CMSE_VENEER_REGION_IN_CODE is not defined.

Change-Id: If32df16bb4e0b3ffac2140acc8d13777e09da152
Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>