fixup! Platform: Add support for new SOCs: nRF5340 and nRF9160
nrf: Extend plat_test.c to implement functions needed for PSA tests

Add timer and memory funtionality.

The PSA_API_TEST_SCRATCH_AREA is used by psa-arch-tests for writing
configurations across resets. Currently no retention is turned on,
since power is assumed to always be on. The area is wiped on
power-on reset.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>